added info embed
